feat(settings): replace flat settings list with drill-in sub-pages
